+++ example/lib/main.dart
@@ -6,285 +6,126 @@ import 'package:camerawesome/pigeon.dart';
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 import 'utils/file_utils.dart';
-import 'package:camerawesome/widgets/camera_with_photo_grid.dart';
-import 'package:camerawesome/models/photo_session.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(const CameraAwesomeApp());
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+class CameraAwesomeApp extends StatelessWidget {
+  const CameraAwesomeApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Camera Awesome Photo Grid Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-        useMaterial3: true,
-      ),
-      home: const HomeScreen(),
+    return const MaterialApp(
+      title: 'camerAwesome',
+      home: CameraPage(),
     );
   }
 }
 
-class HomeScreen extends StatefulWidget {
-  const HomeScreen({super.key});
-
-  @override
-  State<HomeScreen> createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends State<HomeScreen> {
-  final PhotoSession _photoSession = PhotoSession();
-  List<String> _completedPhotos = [];
+class CameraPage extends StatelessWidget {
+  const CameraPage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Camera Awesome Photo Grid'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Card(
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    const Text(
-                      'Photo Session',
-                      style: TextStyle(
-                        fontSize: 20,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    AnimatedBuilder(
-                      animation: _photoSession,
-                      builder: (context, child) {
-                        return Text(
-                          'Current session: ${_photoSession.photoCount} photos',
-                          style: const TextStyle(fontSize: 16),
-                        );
-                      },
-                    ),
-                    const SizedBox(height: 16),
-                    ElevatedButton.icon(
-                      onPressed: () => _openCamera(),
-                      icon: const Icon(Icons.camera_alt),
-                      label: const Text('Open Camera'),
-                      style: ElevatedButton.styleFrom(
-                        padding: const EdgeInsets.all(16),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            
-            const SizedBox(height: 16),
-            
-            if (_completedPhotos.isNotEmpty) ...[
-              Card(
-                child: Padding(
-                  padding: const EdgeInsets.all(16.0),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const Text(
-                        'Last Completed Session',
-                        style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        'Captured ${_completedPhotos.length} photos',
-                        style: const TextStyle(fontSize: 16),
-                      ),
-                      const SizedBox(height: 16),
-                      SizedBox(
-                        height: 100,
-                        child: ListView.builder(
-                          scrollDirection: Axis.horizontal,
-                          itemCount: _completedPhotos.length,
-                          itemBuilder: (context, index) {
-                            return Padding(
-                              padding: const EdgeInsets.only(right: 8.0),
-                              child: ClipRRect(
-                                borderRadius: BorderRadius.circular(8),
-                                child: Image.file(
-                                  File(_completedPhotos[index]),
-                                  width: 80,
-                                  height: 80,
-                                  fit: BoxFit.cover,
-                                  errorBuilder: (context, error, stackTrace) {
-                                    return Container(
-                                      width: 80,
-                                      height: 80,
-                                      color: Colors.grey[300],
-                                      child: const Icon(Icons.image),
-                                    );
-                                  },
-                                ),
-                              ),
-                            );
-                          },
-                        ),
-                      ),
-                      const SizedBox(height: 16),
-                      ExpansionTile(
-                        title: const Text(
-                          'View Image Paths',
-                          style: TextStyle(fontSize: 14, fontWeight: FontWeight.w500),
-                        ),
-                        children: [
-                          Container(
-                            width: double.infinity,
-                            padding: const EdgeInsets.all(12),
-                            decoration: BoxDecoration(
-                              color: Colors.grey[100],
-                              borderRadius: BorderRadius.circular(8),
-                            ),
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: _completedPhotos.asMap().entries.map((entry) {
-                                int index = entry.key;
-                                String path = entry.value;
-                                return Padding(
-                                  padding: const EdgeInsets.only(bottom: 4.0),
-                                  child: SelectableText(
-                                    '${index + 1}. $path',
-                                    style: const TextStyle(
-                                      fontSize: 11,
-                                      fontFamily: 'monospace',
-                                    ),
-                                  ),
-                                );
-                              }).toList(),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-            
-            const Spacer(),
-            
-            const Card(
-              child: Padding(
-                padding: EdgeInsets.all(16.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      'How to use:',
-                      style: TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    SizedBox(height: 8),
-                    Text('1. Tap "Open Camera" to start taking photos'),
-                    Text('2. Tap the capture button to take photos'),
-                    Text('3. Photos appear in the bottom grid'),
-                    Text('4. Tap X on thumbnails to remove photos'),
-                    Text('5. Tap the check mark when done'),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _openCamera() {
-    Navigator.of(context).push(
-      MaterialPageRoute(
-        builder: (context) => CameraWithPhotoGrid(
-          photoSession: _photoSession,
-          onPhotoTaken: (photoPath) {
-            print('Photo taken: $photoPath');
-          },
-          onSessionComplete: (photoPaths) {
-            setState(() {
-              _completedPhotos = photoPaths;
-            });
-            Navigator.of(context).pop();
-            
-            // Print paths to console for debugging
-            print('=== Session Complete ===');
-            print('Total photos: ${photoPaths.length}');
-            for (int i = 0; i < photoPaths.length; i++) {
-              print('Photo ${i + 1}: ${photoPaths[i]}');
+      body: Container(
+        color: Colors.white,
+        child: CameraAwesomeBuilder.awesome(
+          onMediaCaptureEvent: (event) {
+            switch ((event.status, event.isPicture, event.isVideo)) {
+              case (MediaCaptureStatus.capturing, true, false):
+                debugPrint('Capturing picture...');
+              case (MediaCaptureStatus.success, true, false):
+                event.captureRequest.when(
+                  single: (single) {
+                    debugPrint('Picture saved: ${single.file?.path}');
+                  },
+                  multiple: (multiple) {
+                    multiple.fileBySensor.forEach((key, value) {
+                      debugPrint('multiple image taken: $key ${value?.path}');
+                    });
+                  },
+                );
+              case (MediaCaptureStatus.failure, true, false):
+                debugPrint('Failed to capture picture: ${event.exception}');
+              case (MediaCaptureStatus.capturing, false, true):
+                debugPrint('Capturing video...');
+              case (MediaCaptureStatus.success, false, true):
+                event.captureRequest.when(
+                  single: (single) {
+                    debugPrint('Video saved: ${single.file?.path}');
+                  },
+                  multiple: (multiple) {
+                    multiple.fileBySensor.forEach((key, value) {
+                      debugPrint('multiple video taken: $key ${value?.path}');
+                    });
+                  },
+                );
+              case (MediaCaptureStatus.failure, false, true):
+                debugPrint('Failed to capture video: ${event.exception}');
+              default:
+                debugPrint('Unknown event: $event');
             }
-            print('=====================');
-            
-            // Show completion dialog with paths
-            showDialog(
-              context: context,
-              builder: (context) => AlertDialog(
-                title: const Text('Session Complete'),
-                content: SizedBox(
-                  width: double.maxFinite,
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text('Captured ${photoPaths.length} photos successfully!'),
-                      const SizedBox(height: 16),
-                      const Text(
-                        'Image Paths:',
-                        style: TextStyle(fontWeight: FontWeight.bold),
-                      ),
-                      const SizedBox(height: 8),
-                      Container(
-                        constraints: const BoxConstraints(maxHeight: 200),
-                        child: SingleChildScrollView(
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: photoPaths.asMap().entries.map((entry) {
-                              int index = entry.key;
-                              String path = entry.value;
-                              return Padding(
-                                padding: const EdgeInsets.only(bottom: 4.0),
-                                child: Text(
-                                  '${index + 1}. ${path.split('/').last}',
-                                  style: const TextStyle(
-                                    fontSize: 12,
-                                    fontFamily: 'monospace',
-                                  ),
-                                ),
-                              );
-                            }).toList(),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-                actions: [
-                  TextButton(
-                    onPressed: () => Navigator.of(context).pop(),
-                    child: const Text('OK'),
-                  ),
-                ],
+          },
+          saveConfig: SaveConfig.photoAndVideo(
+            initialCaptureMode: CaptureMode.photo,
+            photoPathBuilder: (sensors) async {
+              final Directory extDir = await getTemporaryDirectory();
+              final testDir = await Directory(
+                '${extDir.path}/camerawesome',
+              ).create(recursive: true);
+              if (sensors.length == 1) {
+                final String filePath =
+                    '${testDir.path}/${DateTime.now().millisecondsSinceEpoch}.jpg';
+                return SingleCaptureRequest(filePath, sensors.first);
+              }
+              // Separate pictures taken with front and back camera
+              return MultipleCaptureRequest(
+                {
+                  for (final sensor in sensors)
+                    sensor:
+                        '${testDir.path}/${sensor.position == SensorPosition.front ? 'front_' : "back_"}${DateTime.now().millisecondsSinceEpoch}.jpg',
+                },
+              );
+            },
+            videoOptions: VideoOptions(
+              enableAudio: true,
+              ios: CupertinoVideoOptions(
+                fps: 10,
               ),
+              android: AndroidVideoOptions(
+                bitrate: 6000000,
+                fallbackStrategy: QualityFallbackStrategy.lower,
+              ),
+            ),
+            exifPreferences: ExifPreferences(saveGPSLocation: true),
+          ),
+          sensorConfig: SensorConfig.single(
+            sensor: Sensor.position(SensorPosition.back),
+            flashMode: FlashMode.auto,
+            aspectRatio: CameraAspectRatios.ratio_4_3,
+            zoom: 0.0,
+          ),
+          enablePhysicalButton: true,
+          // filter: AwesomeFilter.AddictiveRed,
+          previewAlignment: Alignment.center,
+          previewFit: CameraPreviewFit.contain,
+          onMediaTap: (mediaCapture) {
+            mediaCapture.captureRequest.when(
+              single: (single) {
+                debugPrint('single: ${single.file?.path}');
+                single.file?.open();
+              },
+              multiple: (multiple) {
+                multiple.fileBySensor.forEach((key, value) {
+                  debugPrint('multiple file taken: $key ${value?.path}');
+                  value?.open();
+                });
+              },
             );
           },
+          availableFilters: awesomePresetFiltersList,
         ),
       ),
     );
